Optimise memory usage (#70)
* Optimise memory use

* Clean up opt code

Remove `SLP_TEST_MODE`

* Employ memory use for parse limit

* Handle fast analysis

+++ smt-log-parser/src/display_with.rs
@@ -216,7 +216,7 @@ mod private {
         pub(super) fn children(&self) -> &'a [TermIdx] {
             self.children
         }
-        pub(super) fn find_quant(&self, idx: &mut usize) -> Option<&Quantifier> {
+        pub(super) fn find_quant(&self, idx: &mut u32) -> Option<&Quantifier> {
             self.quant
                 .iter()
                 .find(|q| {
@@ -469,7 +469,8 @@ impl<'a, 'b> DisplayWithCtxt<DisplayCtxt<'b>, DisplayData<'b>> for &'a TermKind 
         match *self {
             TermKind::Var(mut idx) => {
                 let vars = data.find_quant(&mut idx).and_then(|q| q.vars.as_ref());
-                let name = VarNames::get_name(&ctxt.parser.strings, vars, idx, &ctxt.config);
+                let name =
+                    VarNames::get_name(&ctxt.parser.strings, vars, idx as usize, &ctxt.config);
                 write!(f, "{name}")
             }
             TermKind::App(name) => {
@@ -670,10 +671,11 @@ impl<'a> DisplayWithCtxt<DisplayCtxt<'a>, DisplayData<'a>> for &'a Quantifier {
                     let name = VarNames::get_name(
                         &ctxt.parser.strings,
                         self.vars.as_ref(),
-                        idx,
+                        idx as usize,
                         &ctxt.config,
                     );
-                    let ty = VarNames::get_type(&ctxt.parser.strings, self.vars.as_ref(), idx);
+                    let ty =
+                        VarNames::get_type(&ctxt.parser.strings, self.vars.as_ref(), idx as usize);
                     (idx != 0, name, ty)
                 });
                 let (body, patterns) = data.children().split_last().unwrap();
